requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.4
idna==3.3
jsonify==0.5
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
